netlink: basic test
Because netlink requires sudo, create a separate test binary for it.
The netlink test will run with sudo, called from the test target in the Makefile.
Because this binary is out-of-crate, the listen() method on Backend must be made world-public.
Take care not to call both portus::start() and listen().