ccp_dctcp
An implementation of DCTCP. Very similar to Reno, except a check for ECN
marks and a corresponding tracking of alpha, an ewma-fraction of the
number of marked packets in the last window, is added.

When there are marked packets, the CWND is reduced in proportion to the
fraction of marked packets, rather than Reno's halving.