ccp_bbr
The basic BBR algorithm attempts to find the bottleneck rate.

Send in pulses
--------------
1. For the first RTT, send at 150% the bottleneck rate estimate
2. For the second RTT, send at 50% the bottleneck rate estimate
3. For the next 6 RTTs, send at the bottleneck rate estimate

Gather meaasurements after each phase. If the RcvRate is greater than
the bottleneck rate estimate, set the bottleneck rate estimate
to 0.95 * RcvRate.

BBR is a rate control algorithm. Set the Cwnd to correspond to the
up-pulse rate and a recent RTT.