lang: support +infinity as immediate value
When calculating a minimum, it can be useful for the default to be an
arbitrarily large value. Call this large value "+infinity"

e.g.:
(def (min_rtt +infinity))
(bind Flow.min_rtt (min Flow.min_rtt Rtt))

The immediate value is encoded in the 8 bits as follows:
------------------------
| ImmNum, 2 | Value, 6 |
|    00     |  111111  |
------------------------

The fold API has changed (new immediate), so bump the version to 0.1.2.