lang: Specify events
Change the syntax of the language to encompass control patterns, with an
event-handler-like syntax. Events are specified using the `when` keyword
with one expression specifying whether the event should be evaluated
and any number of subsequent expressions specifying a computation to run
if that event is triggered.

For example:
(def (foo 0))
(when true
    (...)
)
(when (> Ns 1000)
    (:= Rate 1000000)
)
(when (> Ns 2000)
    (:= Rate 2000000)
)

Only one event is evaluated per ccp invocation. This will be fixed in a
subsequent commit.