datapath: Use tuple vector for Scope
Since the number of items in a Scope is limited, it is simpler to use
a vector of tuples instead of a HashMap. This has the added benefit of
outputting registers in sorted order.

If warranted in the future, it is possible to speed up lookup and
insertion using binary search, but again since the number of items
is limited, this is probably unnecessary.