libccp_integration: restructure to separate tests
This commit modifies the structure of the libccp integration tests so
they are run independently, without state from previous tests
interfering with a test. Now, the mock datapath is started independently
for each test. Adding a new integration test involves adding a Test
struct in libccp_integration/src/lib.rs, along with the Impl and CongAlg
trait, that include a install and checker function, and adding code
in libccp_integration/src/bin/test_ccp.rs to run it.