portus: Set variables in scope before installation in datapath
This commit requires users pass in an Option<&[(&str, u32}]> to the
install function, with updates to variables defined in the fast path
program (or None). This is useful for installing the program with
preset values for variables computed in userspace. This commit updates
ccp_generic_cong_avoid for the new API, and adds a new integration test
for this functionality.