lang: Migrate to nom 4.0.0
As a workaround to https://github.com/Geal/nom/issues/544, migrate to
nom 4 to ensure that the verbose-errors feature becomes additive and
therefore portus compiles when used as a dependency.

There were two classes of changes to the parser structure:
1. Error handling, as predicted and outlined here:
https://github.com/Geal/nom/blob/master/doc/upgrading_to_nom_4.md#replacing-parser-result-matchers

2. Migration to "CompleteByteSlice".
This was predicted in the migration notes:
https://github.com/Geal/nom/blob/master/doc/upgrading_to_nom_4.md#dealing-with-incomplete-usage
but caused underlying errors as reported here:
https://github.com/Geal/nom/issues/790#issuecomment-397883681

To address this, shadow nom's `named!` macro with our own,
`named_complete!`, which replaces the types appropriately. This is the
solution proposed here:
https://github.com/Geal/nom/issues/795#issuecomment-398451738