portus api: deduplicate programs across flows
This commit modifies the CCP API to expect users to install all programs
in the new init_programs function, when portus starts running. To make a
flow use a particular program, there is a new set_program function; when
installing, users also specify a name for each program. In the datapath,
now, programs need not be duplicated across flows. This also bumps the
libccp submodule version to match the corresponding changes in libccp.