Appease rustfmt and clippy
There are several clippy warnings we will ignore for now.

"methods called `new` usually return `Self`"
Triggers on IPC constructors. We are returning Result<Self> in these cases.

"very complex type used. Consider factoring parts into `type`
definitions"
This one instance is not worth a bespoke type.

"transmute from a reference to a reference"
In the serialization code. Clippy's suggested fix triggers this lint:
https://rust-lang-nursery.github.io/rust-clippy/v0.0.212/index.html#cast_ptr_alignment
which is worse in severity.

"this argument is passed by value, but not consumed in the function
body"
ipc_valid(): clap calls this function; the signature is fixed.
continue_listening: this is ok.
cfg: the API might change in the future
alg: new_flow() in CongAlg may change to take &mut self, so we should
avoid &U.