Msg: remove lifetime specifier on Other msg type
The Other variant of Msg was previously just a wrapper around RawMsg,
which must have the same lifetime as receive buf, because it has a
direct slice into it. All of the other message variants copy the
contents of RawMsg into their own struct, allowing them to own the data,
meaning they don't need lifetime specifiers. So the only reason we have
a lifetime specifier is for the other type, which itself is only used in
tests and benchmarks.

We can get rid of this by creating a wrapper Other message type that
basically mimics RawMsg, but copies the contents.
This is actually more realistic for the benchmark anyway, because it now
more closely mimics the behavior of the real message types we use (i.e.
it makes its own copy of the message).